Fixes an android crash
Specifically, if you have a view with multiple tabs and you are on a
tab that is not the tab containing the GridView, and you use the App
Switcher to switch to another app and then BACK to your app, it will
crash because owner.items is undefined. Switching the apps appears to
lose the context if the tab you are on is not the tab containing the
Grid View. So, this is a simple check to avoid the crash.